Admin product delete feature implemented

+++ client/src/pages/admin-view/products.jsx
@@ -9,7 +9,7 @@ import {
   SheetTitle,
 } from "@/components/ui/sheet";
 import { addProductFormElements } from "@/config";
-import { addNewProduct, fetchAllProducts } from "@/store/admin/products-slice";
+import { addNewProduct, deleteProduct, fetchAllProducts } from "@/store/admin/products-slice";
 import React, { Fragment, useEffect, useState } from "react";
 import toast from "react-hot-toast";
 import { useDispatch, useSelector } from "react-redux";
@@ -70,6 +70,8 @@ export const AdminProducts = () => {
   }
 
   function handleDelete(getCurrentProductId) {
+    const check= confirm("Do you want to delete this item?");
+    if(check!=true) return;
     dispatch(deleteProduct(getCurrentProductId)).then((data) => {
       if (data?.payload?.success) {
         dispatch(fetchAllProducts());
